Moved code out of main.py
Migrated more code out of main.py and into other files.

Solving algorithms now in solve.py, rubikCube class now in cube.py
